feat: styles are restructured
