+++ src/common/models/index.ts
@@ -1,1 +1,3 @@
 export * from './UserData';
+export * from './message';
+export * from './roomdata';
