+++ src/common/models/message.ts
@@ -1,0 +1,6 @@
+export interface Message {
+  username: string;
+  message: string;
+  date: number;
+  sprite: string;
+}
