+++ src/common/models/roomdata.ts
@@ -1,0 +1,5 @@
+export interface RoomData {
+  users: [];
+  private?: boolean;
+  password?: string;
+}
