+++ src/common/styles/SpriteContainer.ts
