+++ src/features/authenticate/styles/AnimSprite.ts
