+++ src/features/game/components/GameChat.tsx
@@ -3,8 +3,8 @@ import { useSelector, shallowEqual, useDispatch } from 'react-redux';
 import { nanoid } from 'nanoid';
 import convertToDate from '../functions/ConvertToDate';
 import switchSpriteSheet from '../../../common/functions/SwitchSpriteSheet';
-import { Message } from '../../../types/AppTypes';
 import { socket } from '../../../service/socket';
+import { Message } from '../../../common/models';
 import { RootState } from '../../../app/redux';
 import { disableKeyBoard } from '../../../app/redux/GameSlice';
 import useChatScroll from '../hooks/UseChatScroll';
