+++ src/features/game/styles/GameWrapper.ts
@@ -188,21 +188,4 @@ const GameWrapper = styled.main`
   }
 `;
 
-export const SinglePlayerWrapper = styled.main`
-  width: 100vw;
-  height: 100vh;
-  .single-footer {
-    width: 100%;
-    height: 35vh;
-    background: white;
-    overflow-y: scroll;
-  }
-  .game-description {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-`;
-
 export default GameWrapper;
