+++ src/features/user-profile/components/SpriteRadio.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import switchSpriteSheet from '../../../common/functions/SwitchSpriteSheet';
-import SpriteContainer from '../../../styles/SpriteContainer';
+import SpriteContainer from '../../../common/styles/SpriteContainer';
 
 type SpriteProps = {
   newSprite: string;
